+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 
+## 0.12.1 - 2026-08-12
+
+### Fixed
+- Stop conflicts banner from showing with no conflicts
+- Pass return_to on logout link
+
+
+
 ## 0.12.0 - 2026-08-12
 
 ### Added